requirements.txt
# ---------------------------
# Core
# ---------------------------
numpy==1.26.4
pillow>=10.0
tqdm>=4.66
requests>=2.32
rich>=13.7

# ---------------------------
# HF / Diffusers stack
# ---------------------------
diffusers==0.33.0
transformers==4.40.0
accelerate==0.30.0
safetensors>=0.4.0
huggingface_hub==0.24.7

# ---------------------------
# Video export
# ---------------------------
imageio>=2.34
imageio-ffmpeg>=0.5
av>=12.0.0

# ---------------------------
# CLI / Config
# ---------------------------
pydantic>=2.7
python-dotenv>=1.0
typer>=0.12

# ---------------------------
# CV / Pose / FaceID
# ---------------------------
opencv-python==4.10.0.84
insightface==0.7.3
controlnet-aux>=0.0.9

PySide6==6.7.2
pyinstaller==6.10.0